Fix braided pipe support when other mods add pipes
Previously, when braided pipes was enabled, any third-party pipes
that did not specify a connection category would misbehave. In
particular, the color-coded pipes mod fell over spectacularly if
colored and uncolored pipes were ever allowed to mix.

This changes pyindustry to assign its default connection category
to all `pipe` and `pipe-to-ground` instances when they have not
set their own connection category. This causes all such pipes to
behave simply as normal pipes W.R.T. the overall connection scheme.

This is kind of a sledgehammer, but it is also the absolute fallback,
given where it ends up being called. For specific mods other
connection plans might make sense, but as long as their connection
category is set before `data-final-fixes`, this code will ignore
them and leave them as-is.